Moved auto command to appropriate directory, saving wip for driveToDistance auto command.
