+++ src/main/java/frc/robot/Constants.java
@@ -10,6 +10,7 @@ import edu.wpi.first.math.geometry.Translation3d;
 import edu.wpi.first.math.util.Units;
 import swervelib.math.Matter;
 
+import static edu.wpi.first.units.Units.Feet;
 import static edu.wpi.first.units.Units.Meters;
 
 
@@ -35,8 +36,8 @@ public final class Constants {
         public static final PIDConstants AUTO_TRANSLATION_PID = new PIDConstants(0.05, 0.0, 0.0);
         public static final PIDConstants AUTO_ANGLE_PID = new PIDConstants(0.05, 0.0, 0.0);
         // Assume the below is in feet
-        public static final double DRIVE_DISTANCE = 6.0; 
-        public static final double DRIVE_VELOCITY = 0.8;
+        public static final double DRIVE_DISTANCE = Meters.convertFrom(2.0, Feet);
+        public static final double DRIVE_VELOCITY = 1.0;
     }
