+++ src/main/java/frc/robot/commands/auto/MoveForwardAuto.java
@@ -1,7 +1,6 @@
-package frc.robot.commands.actions;
+package frc.robot.commands.auto;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.math.geometry.Translation2d;
 import frc.robot.subsystems.SwerveSubsystem;
 
 
@@ -12,16 +11,17 @@ public class MoveForwardAuto extends Command {
      * Robot starting line is 7 ft, 4 in. Or 224 cm from the reef. 
      */
 
-    private final SwerveSubsystem driveTrain;
-    private boolean isFinished = false;
-    private final double driveDistance;
+//    private final SwerveSubsystem driveTrain;
+//    private boolean isFinished = false;
+//    private final double driveDistance;
+//    private double distance;
 
-    @Override
+
     public MoveForwardAuto(SwerveSubsystem driveTrain, double driveDistance) {
-        this.driveTrain = driveTrain;
-        this.distance = distance;
-        this.currentDist = driveTrain.
-        addRequirements(SwerveSubsystem);
+//        this.driveTrain = driveTrain;
+//        this.distance = distance;
+//        this.currentDist = driveTrain.
+//        addRequirements(driveTrain);
     }
 
     @Override
@@ -30,7 +30,6 @@ public class MoveForwardAuto extends Command {
     @Override
     public void execute() {
         // have current dist, have desired dist, output a velocity output
-        driveTrain.drive(this.driveVel, )
     }
 
     @Override
